pyproject.toml
[project]
name = "axis_explorer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mido>=1.3.2",
    "pillow>=11.0.0",
    "python-rtmidi>=1.5.8",
    "streamdeck>=0.9.5",
]
